// client/typescript-grpc/gen/eventsdb.ts
export interface CreateEventRequest {
  source: string;
  type: string;
  subject: string;
  data: Uint8Array;
}

/** The response message containing the greetings */
export interface CreateEventReply {
  id: number;
}

export interface GetEventByIDRequest {
  id: number;
}

export interface Event {
  id: number;
  source: string;
  type: string;
  subject: string;
  time: string;
  data: Uint8Array;
}

export interface StreamEventsFromSubjectRequest {
  subject: string;
  type?: string | undefined;
  fromId?: number | undefined;
  recursive?: boolean | undefined;
}

export interface StreamEventsFromSubjectReply {
  events: Event[];
}

function createBaseCreateEventRequest(): CreateEventRequest {
  return { source: "", type: "", subject: "", data: new Uint8Array(0) };
}

export const CreateEventRequest = {
  encode(message: CreateEventRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.source !== "") {
      writer.uint32(10).string(message.source);
    }
    if (message.type !== "") {
      writer.uint32(18).string(message.type);
    }
    if (message.subject !== "") {
      writer.uint32(26).string(message.subject);
    }
    if (message.data.length !== 0) {
      writer.uint32(34).bytes(message.data);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): CreateEventRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateEventRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.source = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.type = reader.string();
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.subject = reader.string();
          continue;
        case 4:
          if (tag !== 34) {
            break;
          }

          message.data = reader.bytes();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateEventRequest {
    return {
      source: isSet(object.source) ? globalThis.String(object.source) : "",
      type: isSet(object.type) ? globalThis.String(object.type) : "",
      subject: isSet(object.subject) ? globalThis.String(object.subject) : "",
      data: isSet(object.data) ? bytesFromBase64(object.data) : new Uint8Array(0),
    };
  },

  toJSON(message: CreateEventRequest): unknown {
    const obj: any = {};
    if (message.source !== "") {
      obj.source = message.source;
    }
    if (message.type !== "") {
      obj.type = message.type;
    }
    if (message.subject !== "") {
      obj.subject = message.subject;
    }
    if (message.data.length !== 0) {
      obj.data = base64FromBytes(message.data);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateEventRequest>, I>>(base?: I): CreateEventRequest {
    return CreateEventRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateEventRequest>, I>>(object: I): CreateEventRequest {
    const message = createBaseCreateEventRequest();
    message.source = object.source ?? "";
    message.type = object.type ?? "";
    message.subject = object.subject ?? "";
    message.data = object.data ?? new Uint8Array(0);
    return message;
  },
};

function createBaseCreateEventReply(): CreateEventReply {
  return { id: 0 };
}

export const CreateEventReply = {
  encode(message: CreateEventReply, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.id !== 0) {
      writer.uint32(8).int64(message.id);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): CreateEventReply {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateEventReply();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.id = longToNumber(reader.int64() as Long);
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateEventReply {
    return { id: isSet(object.id) ? globalThis.Number(object.id) : 0 };
  },

  toJSON(message: CreateEventReply): unknown {
    const obj: any = {};
    if (message.id !== 0) {
      obj.id = Math.round(message.id);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateEventReply>, I>>(base?: I): CreateEventReply {
    return CreateEventReply.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateEventReply>, I>>(object: I): CreateEventReply {
    const message = createBaseCreateEventReply();
    message.id = object.id ?? 0;
    return message;
  },
};

function createBaseGetEventByIDRequest(): GetEventByIDRequest {
  return { id: 0 };
}

export const GetEventByIDRequest = {
  encode(message: GetEventByIDRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.id !== 0) {
      writer.uint32(8).int64(message.id);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): GetEventByIDRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetEventByIDRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.id = longToNumber(reader.int64() as Long);
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetEventByIDRequest {
    return { id: isSet(object.id) ? globalThis.Number(object.id) : 0 };
  },

  toJSON(message: GetEventByIDRequest): unknown {
    const obj: any = {};
    if (message.id !== 0) {
      obj.id = Math.round(message.id);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetEventByIDRequest>, I>>(base?: I): GetEventByIDRequest {
    return GetEventByIDRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetEventByIDRequest>, I>>(object: I): GetEventByIDRequest {
    const message = createBaseGetEventByIDRequest();
    message.id = object.id ?? 0;
    return message;
  },
};
// ...
export const StreamEventsFromSubjectRequest = {
  encode(message: StreamEventsFromSubjectRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.subject !== "") {
      writer.uint32(10).string(message.subject);
    }
    if (message.type !== undefined) {
      writer.uint32(18).string(message.type);
    }
    if (message.fromId !== undefined) {
      writer.uint32(24).int64(message.fromId);
    }
    if (message.recursive !== undefined) {
      writer.uint32(32).bool(message.recursive);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): StreamEventsFromSubjectRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseStreamEventsFromSubjectRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.subject = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.type = reader.string();
          continue;
        case 3:
          if (tag !== 24) {
            break;
          }

          message.fromId = longToNumber(reader.int64() as Long);
          continue;
        case 4:
          if (tag !== 32) {
            break;
          }

          message.recursive = reader.bool();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): StreamEventsFromSubjectRequest {
    return {
      subject: isSet(object.subject) ? globalThis.String(object.subject) : "",
      type: isSet(object.type) ? globalThis.String(object.type) : undefined,
      fromId: isSet(object.fromId) ? globalThis.Number(object.fromId) : undefined,
      recursive: isSet(object.recursive) ? globalThis.Boolean(object.recursive) : undefined,
    };
  },

  toJSON(message: StreamEventsFromSubjectRequest): unknown {
    const obj: any = {};
    if (message.subject !== "") {
      obj.subject = message.subject;
    }
    if (message.type !== undefined) {
      obj.type = message.type;
    }
    if (message.fromId !== undefined) {
      obj.fromId = Math.round(message.fromId);
    }
    if (message.recursive !== undefined) {
      obj.recursive = message.recursive;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<StreamEventsFromSubjectRequest>, I>>(base?: I): StreamEventsFromSubjectRequest {
    return StreamEventsFromSubjectRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<StreamEventsFromSubjectRequest>, I>>(
    object: I,
  ): StreamEventsFromSubjectRequest {
    const message = createBaseStreamEventsFromSubjectRequest();
    message.subject = object.subject ?? "";
    message.type = object.type ?? undefined;
    message.fromId = object.fromId ?? undefined;
    message.recursive = object.recursive ?? undefined;
    return message;
  },
};
// ...
export interface EventsDB {
  /** Sends a greeting */
  CreateEvent(request: CreateEventRequest): Promise<CreateEventReply>;
  GetEventByID(request: GetEventByIDRequest): Promise<Event>;
  StreamEventsFromSubject(request: StreamEventsFromSubjectRequest): Observable<StreamEventsFromSubjectReply>;
}

export const EventsDBServiceName = "grpc.EventsDB";
export class EventsDBClientImpl implements EventsDB {
  private readonly rpc: Rpc;
  private readonly service: string;
  constructor(rpc: Rpc, opts?: { service?: string }) {
    this.service = opts?.service || EventsDBServiceName;
    this.rpc = rpc;
    this.CreateEvent = this.CreateEvent.bind(this);
    this.GetEventByID = this.GetEventByID.bind(this);
    this.StreamEventsFromSubject = this.StreamEventsFromSubject.bind(this);
  }
  CreateEvent(request: CreateEventRequest): Promise<CreateEventReply> {
    const data = CreateEventRequest.encode(request).finish();
    const promise = this.rpc.request(this.service, "CreateEvent", data);
    return promise.then((data) => CreateEventReply.decode(_m0.Reader.create(data)));
  }

  GetEventByID(request: GetEventByIDRequest): Promise<Event> {
    const data = GetEventByIDRequest.encode(request).finish();
    const promise = this.rpc.request(this.service, "GetEventByID", data);
    return promise.then((data) => Event.decode(_m0.Reader.create(data)));
  }

  StreamEventsFromSubject(request: StreamEventsFromSubjectRequest): Observable<StreamEventsFromSubjectReply> {
    const data = StreamEventsFromSubjectRequest.encode(request).finish();
    const result = this.rpc.serverStreamingRequest(this.service, "StreamEventsFromSubject", data);
    return result.pipe(map((data) => StreamEventsFromSubjectReply.decode(_m0.Reader.create(data))));
  }
}

interface Rpc {
  request(service: string, method: string, data: Uint8Array): Promise<Uint8Array>;
  clientStreamingRequest(service: string, method: string, data: Observable<Uint8Array>): Promise<Uint8Array>;
  serverStreamingRequest(service: string, method: string, data: Uint8Array): Observable<Uint8Array>;
  bidirectionalStreamingRequest(service: string, method: string, data: Observable<Uint8Array>): Observable<Uint8Array>;
}